QC-35 Made movie listing and upcoming to use similar layout and refactored the backend so it use one movies collection
