+++ cinema-frontend/src/App.js
@@ -18,7 +18,7 @@ import Places from './components/placestogo/Places';
 import Bookings from './components/bookings/Bookings'
 import SearchPage from './components/SearchPage';
 import Payment from './components//bookings/Payment';
-import NewListings from './components/UpComing/NewListings';
+import NewListings from './components/listings/NewListings';
 import Movie from './components/listings/Movie';
 import ContactUs from './components/ContactUs';
