+++ cinema-frontend/src/components/listings/Movie.jsx
@@ -34,6 +34,7 @@ const Movie = () => {
             console.log(error);
         })
     };
+
     if (error) {
         return (
             <p>error</p>
@@ -43,11 +44,11 @@ const Movie = () => {
     } else {
         return (
             <Container>
-                <Row key={movie._id} className="border-bottom py-3">
-                    <img src={`/${movie.poster}`} width="400" height="500" alt={movie.poster} />
-                </Row>
-                <Row>
-                    <Col sm={9} md={10}>
+                <Row className="mt-3">
+                    <Col md={6}>
+                        <Image src={`/${poster}`} alt={poster} fluid />
+                    </Col>
+                    <Col md={6}>
                         <h3>{movie.title}</h3>
                         <p>
                             Actors : {actors.map((actor, index) => (
@@ -59,13 +60,12 @@ const Movie = () => {
                         </p>
                         <div>
                             {showings.map((showing) => (
-                                <Button className="mr-3" variant="primary">{showing}</Button>
+                                <Button className="mr-3 mt-3" variant="primary">{showing}</Button>
                             ))}
                         </div>
                     </Col>
                 </Row>
             </Container>
-
         )
     }
 }
