+++ cinema-frontend/src/components/listings/NewListings.jsx
@@ -1,18 +1,19 @@
 import Container from 'react-bootstrap/Container';
 import Row from 'react-bootstrap/Row';
-import NewMovie from './NewMovie';
 import axios from 'axios';
 import { useState } from 'react';
 import { useEffect } from 'react';
-
+import { Link } from 'react-router-dom';
+import Image from 'react-bootstrap/Image';
+import Col from 'react-bootstrap/Col';
 
 const NewListings = () => {
+
     const [movies, setMovies] = useState([]);
 
     useEffect(() => {
         getMovies();
     }, []);
-
 
     const getMovies = async () => {
         await axios.get("http://localhost:5000/cinema/newmovies").then((response) => {
@@ -26,13 +27,19 @@ const NewListings = () => {
 
     return (
         <Container>
-            <h1>Listings page</h1>
+            <h1>Upcoming listings page</h1>
 
-            <Row className="border-bottom"></Row>
-            {movies.map((movie) => (
-                <NewMovie movie={movie} />
-            ))}
+            <Row xs={1} sm={3} lg={4}>
+                {movies.map((movie, i) => (
+                    <Col key={i} className="mt-4">
+                        <Link to={`/movie/${movie.title}`}>
+                            <Image src={movie.poster} fluid />
+                        </Link>
+                    </Col>
+                ))}
+            </Row>
         </Container>
+
     )
 }
