+++ cinema-frontend/src/components/listings/NewMovie.jsx
